Display more subfields in titles
Moving away from an "only abnp" scheme to an "exclude certain
things, but try to preserve the trailing punctuation where
important" scheme.

We omit the 245$c completely, and we either omit the entire 245$h or
we include just the trailing punctuation.

Initial testing shows that this gives better results in our data
set.

Signed-off-by: Jeff Godin <jgodin@tadl.org>